requirements.txt
astroid==2.4.2
atari-py==0.2.6
cloudpickle==1.3.0
colorama==0.4.3
comtypes==1.1.7
cycler==0.10.0
d3dshot==0.1.5
future==0.18.2
gym==0.17.2
isort==4.3.21
joblib==0.16.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.2
mccabe==0.6.1
MouseInfo==0.1.3
numpy==1.19.0
opencv-python==4.2.0.34
pandas==1.0.5
Pillow==7.1.2
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
pyglet==1.5.0
pylint==2.5.3
PyMsgBox==1.0.8
pyparsing==2.4.7
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2020.1
pywin32==228
scipy==1.5.0
six==1.15.0
stable-baselines==2.10.0
toml==0.10.1
wrapt==1.12.1
